#1360 merge SelenideFieldDecorator with SelenidePageFactory
... to make it easier to customize both of them.

These two classes did use each other anyway
