+++ src/main/java/com/codeborne/selenide/impl/SelenidePageFactory.java
@@ -1,22 +1,34 @@
 package com.codeborne.selenide.impl;
 
 import com.codeborne.selenide.Driver;
+import com.codeborne.selenide.ElementsCollection;
 import com.codeborne.selenide.ElementsContainer;
 import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.ex.PageObjectException;
 import org.openqa.selenium.By;
 import org.openqa.selenium.SearchContext;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.FindBys;
 import org.openqa.selenium.support.pagefactory.Annotations;
+import org.openqa.selenium.support.pagefactory.DefaultElementLocatorFactory;
+import org.openqa.selenium.support.pagefactory.DefaultFieldDecorator;
 import org.openqa.selenium.support.pagefactory.FieldDecorator;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
+import java.lang.reflect.TypeVariable;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * Factory class to make using Page Objects simpler and easier.
@@ -25,6 +37,8 @@ import java.lang.reflect.Type;
  */
 @ParametersAreNonnullByDefault
 public class SelenidePageFactory implements PageObjectFactory {
+  private static final Logger logger = LoggerFactory.getLogger(SelenidePageFactory.class);
+
   @Override
   @CheckReturnValue
   @Nonnull
@@ -44,37 +58,31 @@ public class SelenidePageFactory implements PageObjectFactory {
   @Nonnull
   public <PageObjectClass, T extends PageObjectClass> PageObjectClass page(Driver driver, T pageObject) {
     Type[] types = pageObject.getClass().getGenericInterfaces();
-    initElements(driver, decorator(driver, driver.getWebDriver()), pageObject, types);
+    initElements(driver, driver.getWebDriver(), pageObject, types);
     return pageObject;
-  }
-
-  @CheckReturnValue
-  @Nonnull
-  protected SelenideFieldDecorator decorator(Driver driver, SearchContext searchContext) {
-    return new SelenideFieldDecorator(this, driver, searchContext);
   }
 
   /**
    * Similar to the other "initElements" methods, but takes an {@link FieldDecorator} which is used
    * for decorating each of the fields.
    *
-   * @param decorator the decorator to use
    * @param page      The object to decorate the fields of
    */
-  public void initElements(Driver driver, SelenideFieldDecorator decorator, Object page, Type[] genericTypes) {
+  public void initElements(Driver driver, SearchContext searchContext, Object page, Type[] genericTypes) {
     Class<?> proxyIn = page.getClass();
     while (proxyIn != Object.class) {
-      initFields(driver, decorator, page, proxyIn, genericTypes);
+      initFields(driver, searchContext, page, proxyIn, genericTypes);
       proxyIn = proxyIn.getSuperclass();
     }
   }
 
-  protected void initFields(Driver driver, SelenideFieldDecorator decorator, Object page, Class<?> proxyIn, Type[] genericTypes) {
+  protected void initFields(Driver driver, SearchContext searchContext,
+                            Object page, Class<?> proxyIn, Type[] genericTypes) {
     Field[] fields = proxyIn.getDeclaredFields();
     for (Field field : fields) {
       if (!isInitialized(page, field)) {
         By selector = findSelector(driver, field);
-        Object value = decorator.decorate(page.getClass().getClassLoader(), field, selector, genericTypes);
+        Object value = decorate(page.getClass().getClassLoader(), driver, searchContext, field, selector, genericTypes);
         if (value != null) {
           setFieldValue(page, field, value);
         }
@@ -146,7 +154,104 @@ public class SelenidePageFactory implements PageObjectFactory {
     Constructor<?> constructor = type.getDeclaredConstructor();
     constructor.setAccessible(true);
     ElementsContainer result = (ElementsContainer) constructor.newInstance();
-    initElements(driver, decorator(driver, self), result, genericTypes);
+    initElements(driver, self, result, genericTypes);
     return result;
+  }
+
+
+  @CheckReturnValue
+  @Nullable
+  public final Object decorate(ClassLoader loader,
+                               Driver driver, SearchContext searchContext,
+                               Field field, By selector) {
+    Type[] classGenericTypes = field.getDeclaringClass().getGenericInterfaces();
+    return decorate(loader, driver, searchContext, field, selector, classGenericTypes);
+  }
+
+  @CheckReturnValue
+  @Nullable
+  public Object decorate(ClassLoader loader,
+                         Driver driver, SearchContext searchContext,
+                         Field field, By selector, Type[] genericTypes) {
+    if (ElementsContainer.class.equals(field.getDeclaringClass()) && "self".equals(field.getName())) {
+      if (searchContext instanceof SelenideElement) {
+        return searchContext;
+      }
+      else {
+        logger.warn("Cannot initialize field {}", field);
+        return null;
+      }
+    }
+    if (WebElement.class.isAssignableFrom(field.getType())) {
+      return ElementFinder.wrap(driver, searchContext, selector, 0);
+    }
+    if (ElementsCollection.class.isAssignableFrom(field.getType()) ||
+      isDecoratableList(field, genericTypes, WebElement.class)) {
+      return new ElementsCollection(new BySelectorCollection(driver, searchContext, selector));
+    }
+    else if (ElementsContainer.class.isAssignableFrom(field.getType())) {
+      return createElementsContainer(driver, searchContext, field, selector);
+    }
+    else if (isDecoratableList(field, genericTypes, ElementsContainer.class)) {
+      return createElementsContainerList(driver, searchContext, field, genericTypes, selector);
+    }
+
+    return defaultFieldDecorator(searchContext).decorate(loader, field);
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  protected DefaultFieldDecorator defaultFieldDecorator(SearchContext searchContext) {
+    return new DefaultFieldDecorator(new DefaultElementLocatorFactory(searchContext));
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  protected List<ElementsContainer> createElementsContainerList(Driver driver, SearchContext searchContext,
+                                                                Field field, Type[] genericTypes, By selector) {
+    Class<?> listType = getListGenericType(field, genericTypes);
+    if (listType == null) {
+      throw new IllegalArgumentException("Cannot detect list type for " + field);
+    }
+
+    return new ElementsContainerCollection(this, driver, searchContext, field, listType, genericTypes, selector);
+  }
+
+  @CheckReturnValue
+  protected boolean isDecoratableList(Field field, Type[] genericTypes, Class<?> type) {
+    if (!List.class.isAssignableFrom(field.getType())) {
+      return false;
+    }
+
+    Class<?> listType = getListGenericType(field, genericTypes);
+
+    return listType != null && type.isAssignableFrom(listType)
+      && (field.getAnnotation(FindBy.class) != null || field.getAnnotation(FindBys.class) != null);
+  }
+
+  @CheckReturnValue
+  @Nullable
+  protected Class<?> getListGenericType(Field field, Type[] genericTypes) {
+    Type fieldType = field.getGenericType();
+    if (!(fieldType instanceof ParameterizedType)) return null;
+
+    Type[] actualTypeArguments = ((ParameterizedType) fieldType).getActualTypeArguments();
+    Type firstType = actualTypeArguments[0];
+    if (firstType instanceof TypeVariable) {
+      int indexOfType = indexOf(field.getDeclaringClass(), firstType);
+      return (Class<?>) genericTypes[indexOfType];
+    }
+    else if (firstType instanceof Class) {
+      return (Class<?>) firstType;
+    }
+    throw new IllegalArgumentException("Cannot detect list type of " + field);
+  }
+
+  protected int indexOf(Class<?> klass, Type firstArgument) {
+    Object[] objects = Arrays.stream(klass.getTypeParameters()).toArray();
+    for (int i = 0; i < objects.length; i++) {
+      if (objects[i].equals(firstArgument)) return i;
+    }
+    return -1;
   }
 }
